requirements.txt
discord.py
requests
python-dotenv
audioop-lts
